Add feature to change board size
There's now a button that prompts the user to input a number.
The board will then reset to use that number as its dimensions in terms
of how many 'pixel' divs are inside.

We noticed that every time the board resets, we have to re-query each
'pixel' div to give each of them the ability to change color. So we
separated that action into a dedicated function.